+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [1.0.7](https://github.com/mangata-finance/mangata-SDK/compare/v1.0.8...v1.0.7) (2022-06-24)
+### [1.0.7](https://github.com/mangata-finance/mangata-SDK/compare/v1.0.6...v1.0.7) (2022-06-24)
+
+- fix method wait for new block using subscribe new heads
+- better handling get token balance method
 
 ### [1.0.6](https://github.com/mangata-finance/mangata-SDK/compare/v1.0.5...v1.0.6) (2022-06-19)